+++ CoffeeMaker/src/exceptions/StaleCoffeeException.java
@@ -1,0 +1,8 @@
+package exceptions;
+
+public class StaleCoffeeException extends Exception {
+
+    public StaleCoffeeException(int timeSinceLastBrew) {
+        super("Coffee is stale since " + String.valueOf(timeSinceLastBrew) + " minutes.");
+    }
+}