+++ ExceptionalZoo/src/model/Keeper.java
@@ -1,5 +1,9 @@
 package model;
 
+import exceptions.AllergyException;
+import exceptions.MessyError;
+import exceptions.NotHungry;
+
 import java.util.List;
 
 public class Keeper {
@@ -10,12 +14,12 @@ public class Keeper {
         animalsToFeed = animals;
     }
 
-    public void feed() {
+    public void feed() throws NotHungry, AllergyException {
+        System.out.println("Keeper is feeding the animals.");
         for (Animal animal : animalsToFeed) {
-            if (!animal.isHungry()) {
-                int eatenTimes = animal.eat();
-                System.out.println("Animal has been fed " + eatenTimes);
-            }
+            int eatenTimes = animal.eat();
+            System.out.println("Animal has been fed " + eatenTimes + " times.");
         }
+        throw new MessyError();
     }
 }