+++ Weather/src/exceptions/ImpossibleException.java
@@ -1,0 +1,4 @@
+package exceptions;
+
+public class ImpossibleException extends Exception {
+}