+++ Weather/src/exceptions/ImpossiblePrecipException.java
@@ -1,0 +1,4 @@
+package exceptions;
+
+public class ImpossiblePrecipException extends ImpossibleException {
+}