+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.2] - 2017-11-12
+### Added
+- CoffeeMaker/src/exceptions
+- Weather project
+
+### Changed
+- CoffeeMaker/src/main/Main
+- CoffeeMaker/src/model/CoffeeMaker
+- CoffeeMaker/src/test/CoffeeMakerTest
+- README.md - Table of content
+
 ## [0.0.1] - 2017-10-11
 ### Added
 - CHANGELOG.md file